+++ CHANGELOG.md
@@ -18,6 +18,13 @@ The changelog for **SwifterSwiftUI**. Also see the [releases](https://github.com
 
 ---
 
+## [v0.4.0](https://github.com/asam139/SwifterSwiftUI/releases/tag/0.4.0)
+
+### Added
+
+- **Storage**
+    - Property wrapper to use the user's defaults.
+
 ## [v0.3.0](https://github.com/asam139/SwifterSwiftUI/releases/tag/0.3.0)
 
 ### Added